activity file updated

+++ coa/dal/activity.php
@@ -31,25 +31,10 @@ class activity {
     public function Add($ActivityType,$ActivityDescription)
     {
         global $db;
-		
-        $Sql = 'INSERT INTO '.$this->TableName.' VALUES ('.$ActivityType.','.$ActivityDescription.')';
-        try{
-        $query = $db->prepare($Sql);
-        $query->execute();
-                $result['msg'] = true;
-                $result['data'] = $query->fetchAll(PDO::FETCH_ASSOC);
-        }catch(PDOException $e){
-                $result['msg'] = false;
-                $result['data'] = $e->getMessage();
-        }
-		
-	return $result;
         
-    }
-	
-	public function Update($ActivityType,$ActivityDescription,$ActivityId){
-            global $db;
-            $Sql = 'UPDATE '.$this->TableName.' SET name = '.$ActivityType.', description='.$ActivityDescription.' WHERE id ='.$ActivityId;
+        if($ActivityType != "" && $ActivityDescription != ""){
+		
+            $Sql = 'INSERT INTO '.$this->TableName.' VALUES ('.$ActivityType.','.$ActivityDescription.')';
             try{
             $query = $db->prepare($Sql);
             $query->execute();
@@ -59,6 +44,31 @@ class activity {
                     $result['msg'] = false;
                     $result['data'] = $e->getMessage();
             }
+        }else{
+             $result['msg'] = false;
+             $result['data'] = 'Invalid parameters';
+        }
+        return $result;
+    }
+	
+	public function Update($ActivityType,$ActivityDescription,$ActivityId){
+            global $db;
+            
+            if($ActivityType != "" && $ActivityDescription != "" && $ActivityId != ""){
+                $Sql = 'UPDATE '.$this->TableName.' SET name = '.$ActivityType.', description='.$ActivityDescription.' WHERE id ='.$ActivityId;
+                try{
+                $query = $db->prepare($Sql);
+                $query->execute();
+                        $result['msg'] = true;
+                        $result['data'] = $query->fetchAll(PDO::FETCH_ASSOC);
+                }catch(PDOException $e){
+                        $result['msg'] = false;
+                        $result['data'] = $e->getMessage();
+                }
+            }else{
+                 $result['msg'] = false;
+                $result['data'] = 'Invalid parameters';
+            }
 
             return $result;
 	}
@@ -66,15 +76,20 @@ class activity {
 	public function Delete_record($ActivityId){
             global $db;
 		
-            $Sql = "DELETE FROM " . $this->TableName . " WHERE id =" . $ActivityId;
-            try{
-            $query = $db->prepare($Sql);
-            $query->execute();
-                $result['msg'] = true;
-                $result['data'] = $query->fetchAll(PDO::FETCH_ASSOC);
-            }catch(PDOException $e){
+            if($ActivityId != ""){
+                $Sql = "DELETE FROM " . $this->TableName . " WHERE id =" . $ActivityId;
+                try{
+                $query = $db->prepare($Sql);
+                $query->execute();
+                    $result['msg'] = true;
+                    $result['data'] = $query->fetchAll(PDO::FETCH_ASSOC);
+                }catch(PDOException $e){
+                    $result['msg'] = false;
+                    $result['data'] = $e->getMessage();
+                }
+            }else{
                 $result['msg'] = false;
-                $result['data'] = $e->getMessage();
+                $result['data'] = 'Invalid parameters';
             }
 
             return $result;
